IIIF range folders now display annotation count
