+++ src/pages/images/IIIFManifestOverview/IIIFManifestGrid/IIIFRangeItem.tsx
@@ -27,10 +27,11 @@ export const IIIFRangeItem = (props: IIIFRangeItemProps) => {
             className="scale w-[190px] h-[190px] transition-all drop-shadow-md" />
 
           <IIIFIcon
-            className="iiif-logo text-white transition-all absolute top-5 left-4 size-6" />
+            light
+            className="iiif-logo text-white transition-all absolute top-5 left-4 size-5" />
         </button>
 
-        <div className="absolute bottom-3 px-3 pt-10 pb-3 left-1.5 w-full pointer-events-auto text-white text-sm">
+        <div className="absolute bottom-2.5 px-3 pt-10 pb-3 left-1.5 w-full pointer-events-auto text-white text-sm">
           <MessagesSquare 
             size={18} 
             className="inline align-text-bottom mr-1" />
