+++ src/pages/images/ImagesUtils.tsx
@@ -1,8 +1,15 @@
-import { ArrowDownNarrowWide, ArrowDownWideNarrow, ArrowUpDown, MessagesSquare } from 'lucide-react';
 import Moment from 'react-moment';
+import murmur from 'murmurhash';
+import { CozyRange } from 'cozy-iiif';
 import { W3CAnnotation } from '@annotorious/react';
 import { ColumnSortEvent } from 'primereact/column';
-import { ItemTableRow } from './Types';
+import { AnnotationMap, ItemTableRow } from './Types';
+import { 
+  ArrowDownNarrowWide, 
+  ArrowDownWideNarrow, 
+  ArrowUpDown, 
+  MessagesSquare 
+} from 'lucide-react';
 
 export const TABLE_HEADER_CLASS = 'pl-3 pr-2 whitespace-nowrap text-xs text-muted-foreground font-semibold text-left';
 
@@ -93,3 +100,18 @@ export const ANNOTATIONS_COLUMN_TEMPLATE = (row: ItemTableRow) => (
     </div>
   </div>
 )
+
+export const getAnnotationsInRange = (range: CozyRange, annotations: Record<string, W3CAnnotation[]>): W3CAnnotation[] => {
+  // Canvases directly contained in this range
+  const annotationsOnCanvases = range.canvases.reduce<W3CAnnotation[]>((agg, canvas) => {
+    const id = murmur.v3(canvas.id);
+    return [...agg, ...(annotations[id] || [])];
+  }, []);
+
+  // Subranges
+  const annotationsOnSubRanges = range.ranges.reduce<W3CAnnotation[]>((agg, range) => {
+    return [...agg, ...getAnnotationsInRange(range, annotations)]
+  }, []);
+
+  return [...annotationsOnCanvases, ...annotationsOnSubRanges];
+}
